Fixed Telegram
Intrudced a number of fixes to make the telegram bot work again. First, i implemented the necessary functions in database class to update and retrieve the api key. Second i added the functionality in the settings tab to update the API key if necessary. Third i adapted all uses of the api key to load from the Database from now on.